Moved the functions that take random stuff into utils test
they were not needed in the utils function
